# ishop-api-contract.yaml
openapi: 3.1.0
info:
  title: iShop API
  description: This is the OpenAPI for the backend iShop API
  version: 1.0.0
servers:
  - url: 'http://localhost:3000/api/v1'
paths:
...
  /users/email/{email}:
    get:
      summary: Get user by email
      description: Retrieve a user by their email (requires authentication)
      operationId: getUserByEmail
      tags:
        - Users
      security:
        - basicAuth: [ ]
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
            format: email
      responses:
        '200':
          description: User retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: bearer
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [ user, admin, workers ]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - email
        - role
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        details:
          type: object
          description: Additional error details
      required:
        - error